requirements.txt
# Core Python
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
xarray==2023.12.0
netCDF4==1.6.5
dask[complete]==2023.12.1
polars==0.20.3  # Optional alternative to pandas

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
asyncpg==0.29.0  # Async PostgreSQL driver

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Caching
redis==5.0.1
hiredis==2.3.2  # C parser for Redis

# Task Scheduling
apscheduler==3.10.4
# prefect==2.14.0  # Alternative to Airflow, uncomment if using

# Scientific Computing
scipy==1.11.4

# Geospatial (if needed for domain validation)
geopandas==0.14.1  # Optional
shapely==2.0.2  # Optional

# Configuration
PyYAML==6.0.1
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2  # Optional

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.3  # Retry logic
